Preserve builder state on clean builds
When calling `gradle clean`, the Xtext builder was previously discarded,
which greatly degraded performance in multi-project builds. For each generator
task after a `clean`, the builder und thus the whole index was recreated.

This has been replaced with a smarter up-to-date detection for the builder
classpath. The language setups, default encoding and classpath entry names
and modification timestamps are taken into account to determine whether an
existing builder can be reused.