Reduce usage of internal/discouraged APIs
Manipulating the Xtext classpath shortly before the generator
task is executed was not necessary and would have broken features
like composite builds. Instead, we compute the Xtext version lazily
when the Xtext classpath is first resolved.

The FileResolver was unused, so we could remove another internal API
usage.

The final remaining internal API usage is how we plug the Xtend
sourceSet into the Java sourceSet. This is not easily solved, as
all of Gradle's JVM language plugins use the same pattern. At least
it is therefor unlikely to change.