Fix compatiblity with Android plugin
Packaging options need to be configured before variants are evaluated.

The Xtend classpath inferrer was called to early and overwritten. The inference
mechanism was improved to be more explicit. There is now a chain of inferrers which
is used when the user did not explicitly overwrite the Xtext classpath. This chain
could even be used by other language authors to infer their own language classpath.