Improve version detection regex
Match any kind of version number, no matter how many dashes it contains
or whether it starts with a number or letter. Fixes #81

Only match a few well-known libraries instead of everything that starts
with org.eclipse.xtext, since there are some Xtext example languages that
start with that prefix but don't have the same version as Xtext. Fixes #129