Remove use of afterEvaluate
Use convention mapping instead, which was the better way of wiring
things up until Gradle introduced the Provider API in more recent
versions. Convention mapping is not going away any time soon, since
it is still heavily used in Gradle itself to stay backwards compatible.

Convention mapping is basically a more cumbersome and weakly typed
version of the Provider API, but with the same benefits. Wiring is lazy,
so we no longer depend on point-in-time snapshots of the project's config
like when using afterEvaluate.

Replace the afterEvaluate usage for our source set includes with handling
the includes in our source set implementation itself. This makes it more
reliable for the user, since they will get the correct sources at any given
point in time.