Restore JavaCompile dependency on generator
This partly reverts the last change, since the JavaCompile task seems
to lose the dependencies declared on SourceDirectorySet#java. So we need
to declare an explicit depencency on the generator after all. This is a bug
in Gradle and probably wasn't fixed yet because there are no code generator
tasks in Gradle itself, only compilers like Java/Groovy/Scala.

At least this showed us that we had no integration test coverage for this
task dependency, so I've added a test.