Improve version alignment of orc.eclipse.core.runtime
There are old packages of this dependency out there which are not in
the org.eclipse.platform group, but in org.eclipse.core. Match those
as well and upgrade them to the correct org.eclipse.platform equivalent
to avoid different versions of effectively the same JAR on the classpath.

Fixes #121