Simplify xtextTooling dependency setup
Create the xtextTooling Configuration for each SourceSet right away
rather than lazily. This makes it visible in the output of the `dependencies`
task and allows users to easily customize it in their build scripts, e.g.
when there is another issue like https://github.com/xtext/xtext-gradle-plugin/issues/187

Use lazy callback APIs like `eachDependency` (to align versions for Xtext itself)
and `withDependencies` (to add the BOM when that Xtext version has one) instead of
using the Gradle-interna LazilyInitializedFileCollection.

Load all dependencies, including the Xtext languages into the xtextTooling Configuration
so that all version numbers are aligned. Previously it was possible for the xtextLanguages
Configuration to bring in incompatible versions of libraries like eclipse.core.resources,
which was especially visible when trying to use Xcore with this plugin. Now everything
is aligned with the Xtext BOM and/or our dependency resolution rules.

Remove the complicated "classpathInferrer" logic in favor of this new approach.
This also means that the Xtend plugin is now much simpler, since it only needs to add
the xtend.core dependency to the xtextTooling Configurations and that version will automatically
be aligned by our resolution rules.

This raises the minimum Gradle version to 4.7, which seems very reasonable given that the
latest is 7.2.